Fixup home page path function
