+++ next.config.js
@@ -4,7 +4,7 @@ const nextConfig = {
   experimental: {},
   typescript: {
     // Disable build errors since dev dependencies aren't loaded on prod. Rely on GitHub actions to throw any errors.
-    ignoreBuildErrors: process.env.CI !== 'true',
+    // ignoreBuildErrors: process.env.CI !== 'true',
   },
   images: {
     remotePatterns: [
